Refactor tools drawer
Create version-specific tool drawers
