+++ src/components/v2/V2Project/V2ProjectHeaderActions.tsx
@@ -1,6 +1,6 @@
 import { t, Trans } from '@lingui/macro'
 import { Button, Tooltip } from 'antd'
-import { ProjectToolsDrawer } from 'components/Project/ProjectToolsDrawer/ProjectToolsDrawer'
+import { V2ProjectToolsDrawer } from 'components/v2/V2Project/V2ProjectToolsDrawer/V2ProjectToolsDrawer'
 import { ThemeContext } from 'contexts/themeContext'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 import { useV2ConnectedWalletHasPermission } from 'hooks/v2/contractReader/V2ConnectedWalletHasPermission'
@@ -9,20 +9,12 @@ import { V2OperatorPermission } from 'models/v2/permissions'
 import { useContext, useState } from 'react'
 import { ToolOutlined } from '@ant-design/icons'
 
-import { useAddToBalanceTx } from 'hooks/v2/transactor/AddToBalanceTx'
-import { useTransferProjectOwnershipTx } from 'hooks/v2/transactor/TransferProjectOwnershipTx'
-import { useTransferUnclaimedTokensTx } from 'hooks/v2/transactor/TransferUnclaimedTokensTx'
-import useUserUnclaimedTokenBalance from 'hooks/v2/contractReader/UserUnclaimedTokenBalance'
-import { useDeployProjectPayerTx } from 'hooks/v2/transactor/DeployProjectPayerTx'
-
 import ProjectVersionBadge from 'components/ProjectVersionBadge'
-import { useEditV2ProjectDetailsTx } from 'hooks/v2/transactor/EditV2ProjectDetailsTx'
 
 import V2ReconfigureFundingModalTrigger from './V2ProjectReconfigureModal/V2ReconfigureModalTrigger'
 
 export default function V2ProjectHeaderActions() {
-  const { projectId, tokenSymbol, projectOwnerAddress } =
-    useContext(V2ProjectContext)
+  const { projectId } = useContext(V2ProjectContext)
 
   const {
     theme: { colors },
@@ -35,8 +27,6 @@ export default function V2ProjectHeaderActions() {
   )
 
   const showReconfigureButton = canReconfigure
-
-  const { data: unclaimedTokenBalance } = useUserUnclaimedTokenBalance()
 
   return (
     <div
@@ -58,18 +48,9 @@ export default function V2ProjectHeaderActions() {
           <ProjectVersionBadge versionText="V2" />
         </Tooltip>
       </span>
-      <ProjectToolsDrawer
+      <V2ProjectToolsDrawer
         visible={toolDrawerVisible}
         onClose={() => setToolDrawerVisible(false)}
-        unclaimedTokenBalance={unclaimedTokenBalance}
-        tokenSymbol={tokenSymbol}
-        ownerAddress={projectOwnerAddress}
-        useTransferProjectOwnershipTx={useTransferProjectOwnershipTx}
-        useTransferUnclaimedTokensTx={useTransferUnclaimedTokensTx}
-        useAddToBalanceTx={useAddToBalanceTx}
-        useSetProjectUriTx={() => undefined}
-        useEditV2ProjectDetailsTx={useEditV2ProjectDetailsTx}
-        useDeployProjectPayerTx={useDeployProjectPayerTx}
       />
       <div style={{ display: 'flex' }}>
         <Tooltip title={t`Tools`} placement="bottom">
