+++ src/components/v2/V2Project/V2ProjectToolsDrawer/index.tsx
@@ -1,0 +1,1 @@
+export { V2ProjectToolsDrawer } from './V2ProjectToolsDrawer'
